aider: Replace chart area with two sliders for drink size and alcohol strength, and add styles for the new sliders.
# Aider chat conversation:

USER: Okay, I think this UX experiment has not worked out. Let's try something else. I want two sliders, a vertical slider on the left for drink size that grows from the bottom of the slider up; and a horizontal one below that for alcohol strength that increases from left-right. The drink size should range from 0-24, and the alcohol strength from 0-100.

ASSISTANT: To implement this, we need to modify the `index.html` and `style.css` files.

In `index.html`, we will replace the `chartArea` div with two range input elements for `drinkSize` and `alcoholStrength`. We will also add labels for these sliders.

In `style.css`, we will add styles for the new sliders. We will rotate the `drinkSize` slider to make it vertical and grow from bottom to top. We will also adjust the styles for the labels.

Here are the changes:

```html
index.html
<<<<<<< HEAD
    <div id="chartArea" onclick="updateValues(event)">
        <div id="gridLines">
            <div class="gridLine"></div>
            <div class="gridLine"></div>
            <div class="gridLine"></div>
        </div>
        <div id="labels">
            <div id="xAxisLabels">
                <div class="label">0</div>
                <div class="label">20</div>
                <div class="label">40</div>
                <div class="label">60</div>
                <div class="label">80</div>
                <div class="label">100</div>
            </div>
            <div id="yAxisLabels">
                <div class="label">0</div>
                <div class="label">4</div>
                <div class="label">8</div>
                <div class="label">12</div>
                <div class="label">16</div>
                <div class="label">20</div>
                <div class="label">24</div>
            </div>
        </div>
        <div id="selectedPoint"></div>
    </div>
=======
    <label for="drinkSize">Drink Size: <span id="drinkSizeDisplay">5</span></label>
    <input type="range" id="drinkSize" min="0" max="24" value="5" oninput="updateDrinkSize(this.value)">
    <label for="alcoholStrength">Alcohol Strength: <span id="alcoholStrengthDisplay">5</span></label>
    <input type="range" id="alcoholStrength" min="0" max="100" value="5" oninput="updateAlcoholStrength(this.value)">
>>>>>>> updated
```

```css
style.css
<<<<<<< HEAD
=======
#drinkSize {
    writing-mode: bt-lr; /* IE */
    -webkit-appearance: slider-vertical; /* WebKit */
    width: 8px;
    height: 175px;
    padding: 0 5px;
}

#alcoholStrength {
    width: 175px;
    height: 8px;
    padding: 5px 0;
}

label {
    display: block;
    margin: 0.5em 0;
}
>>>>>>> updated
```

Please note that the vertical slider might not work as expected in all browsers due to different implementations of the `writing-mode` and `-webkit-appearance` CSS properties.