aider: Add a media query for screens smaller than 600px to make the ounces and strength into columns.
# Aider chat conversation:

USER: Since a phone screen is usually in portrait orientation, can we make the ounces and strength into columns?

ASSISTANT: Sure, we can modify the CSS to make the ounces and strength into columns on smaller screens. We will need to modify the `style.css` file.

Here are the changes:

1. Add a media query for screens smaller than 600px. Inside this media query, we will set the display of the `.column` class to `flex` and the flex-direction to `column`. This will stack the buttons vertically on smaller screens.

Here is the *edit block* for this change:

```css
style.css
<<<<<<< HEAD
@media (max-width: 480px) {
    button {
        padding: 1em 2em;
        font-size: 1.5em;
    }

    #alcoholCount {
        font-size: 2em;
    }
}
=======
@media (max-width: 600px) {
    .column {
        display: flex;
        flex-direction: column;
    }

    button {
        padding: 1em 2em;
        font-size: 1.5em;
    }

    #alcoholCount {
        font-size: 2em;
    }
}
>>>>>>> updated
```

Please confirm these changes so I can proceed.